refactor `output_voltage`
